refactor(ocr): share native stack validation
